trk header and offsets in separate header files
